% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{delete_map}
\alias{delete_map}
\title{Erase map from the map server.}
\usage{
delete_map(id, map_data = "inst/extdata")
}
\arguments{
\item{id}{a numeric indicating the primary key (ID) of the database entry
concerning a specific map.}

\item{map_data}{a character string with the path to the folder storing the map data.}
}
\description{
The \code{delete_map} function erases a selected map entry from the
database and the map file from the map folders.
}
