% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{display_map}
\alias{display_map}
\title{Load and plot map file given the database record ID.}
\usage{
display_map(id, map_data = "W:/DANE/Dane_OpenStreetMap/Serwer_map",
  mapdb_name = "MapID")
}
\arguments{
\item{id}{an integer holding the number of the record in the map database.}

\item{map_data}{a character string with the path to the folder storing the map data.}

\item{mapdb_name}{a character string with the name of the map 
database.}
}
\description{
The \code{display_map} function loads the map raster background and plots it
using the R graphics device.
}
\examples{
###########################################################
# Displaying the map database
display_map_info()

# Displaying the chosen map	
display_map(id = 50)
###########################################################
}
