% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{display_map_info}
\alias{display_map_info}
\title{Display the list of available maps.}
\usage{
display_map_info(map_data = "inst/extdata")
}
\arguments{
\item{map_data}{a character string with the path to the folder storing the map data.}
}
\description{
The \code{display_map_info} function fetches all the entries from the MapID database
and prints them.
}
