% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{establish_con}
\alias{establish_con}
\title{Establish connection to the map database (default name 'MapID').}
\usage{
establish_con(map_data = "extdata", mapdb_name = "MapID.db")
}
\arguments{
\item{map_data}{a character string with the path to the folder 
storing the map data, a default value has been set
in the function.}

\item{mapdb_name}{a character string with the name of the map 
database.}
}
\value{
db a SQLiteConnection object.
}
\description{
The \code{establish_con} function checks the existence of the map
database and establishes a connection which can be then passed on
to other functions. The required information is the path to the
directory holding map data and optionally a name of the database, 
if it is different than default.
}
\keyword{internal}
