% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{load_map}
\alias{load_map}
\title{Load and plot map file given the database record.}
\usage{
load_map(mapl, map_data, ext_map)
}
\arguments{
\item{mapl}{a data frame holding one record from the map database.}

\item{map_data}{a character string with the path to the folder storing the map data.}

\item{ext_map}{a spatial object of class 'extent' from the 'raster' 
package holding the data on the extreme coordinates of 
the input object including a preset margin.}
}
\description{
The \code{load_map} function loads the map raster background and plots it
using the R graphics device.
}
\keyword{internal}
