% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{render_map}
\alias{render_map}
\title{Renders required map background using OSM data and mapnik rendering engine.}
\usage{
render_map(map_data, ext_map, options_list = NULL, db)
}
\arguments{
\item{map_data}{a character string with the path to the folder storing the map data.}

\item{ext_map}{a spatial object of class 'extent' from the 'raster' package
holding the data on the extreme coordinates of the input object.}

\item{options_list}{a list of graphical and other parameters used by the rendering engine.}

\item{db}{a SQLiteConnection object that specifies the connection to the database.}
}
\value{
NULL 	there are no values returned by this function to the workspace; the output
				of the function is just the map file written in the /ttemp folder.
}
\description{
The \code{render_map} passes the bounding box and the graphical parameters of the 
required map background to the mapnik rendering engine. The map is rendered by mapnik
based on OpenStreetMap data stored in a local PostgreSQL database.
}
\keyword{internal}
