% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{save_map}
\alias{save_map}
\title{Save the map file to a proper subdirectory and update the database.}
\usage{
save_map(db, init_proj, ext_map, map_data, options_list = NULL)
}
\arguments{
\item{db}{a SQLiteConnection object that specifies the connection to the database.}

\item{init_proj}{a character string that specifies the EPSG code of the geographic
projection of the map file (by default "+init=epsg:4326").}

\item{ext_map}{a spatial object of class 'extent' from the 'raster' package
holding the data on the extreme coordinates of the input object  including margins.}

\item{map_data}{a character string with the path to the folder storing the map data.}

\item{options_list}{a list of graphical and other parameters used by the rendering engine.}
}
\value{
mapl     a data frame holding the last record from the map database that has been 
				entered to the database by the last function run.
}
\description{
The \code{save_map} function renames and copies the map file to the 
appropriate directory after rendering and updates the database with 
the map attributes.
}
\keyword{internal}
