% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funcs.R
\name{validate_mapid}
\alias{validate_mapid}
\title{Check that stored map info is not corrupt.}
\usage{
validate_mapid(map_data = "extdata")
}
\arguments{
\item{map_data}{a character string with the path to the folder storing the map data,
a default value has been set in the function.}
}
\description{
The \code{validate_mapid} function checks if the entries in the MapID database
match the files stored in the map folders and erases the files without a matching
entry; if there are more entries than files, then the missing map files are being
rendered.
}
